# server/swagger.yml
openapi: "3.0.2"
info:
  title: Time2Bee API
  version: "1.0"
servers:
  - url: http://localhost:3000/api
  - url: https://oxypomme.fr/time2bee/api
components:
  schemas:
    Plant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        latin_name:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: object
            properties:
              hex:
                type: string
              french:
                type: string
        flowering:
          type: array
          items:
            type: integer
        garden_spot:
          type: string
        height:
          type: number
          format: float
        honeydrew:
          type: boolean
        pollen:
          type: integer
          format: int32
        nectar:
          type: integer
          format: int32
        creationDate:
          type: string
          format: date-time
        editedDate:
          type: string
          format: date-time
    PlantPayload:
      type: object
      properties:
        name:
          type: string
        latin_name:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        pollen:
          type: integer
          format: int32
        nectar:
          type: integer
          format: int32
    Error:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
paths:
  /plants:
    get:
      description: List all available plants
      tags:
        - Plant
      responses:
        "200":
          description: Returns the list of available plants.
          content:
            "application/json":
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    latin_name:
                      type: string
        "500":
          description: An error occured when fetching plants.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
    post:
      description: Add a new plants
      tags:
        - Plant
      responses:
        "200":
          description: The new plant details
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Plant"
        "409":
          description: Another plant already have the same `name` and `latin_name`
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: An unexpected error happened.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        description: Plant to add
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/PlantPayload"
  /plants/{id}:
    parameters:
      - name: id
        in: path
        description: id of the plant
        required: true
        schema:
          type: string
    get:
      description: Get informations about a plant
      tags:
        - Plant
      responses:
        "200":
          description: Returns the details of a plant.
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plant"
        "404":
          description: No plant have the given `id`
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: An error occured when fetching plants.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      description: Delete informations about a plant
      tags:
        - Plant
      responses:
        "200":
          description: The deletion is successful.
        "404":
          description: No plant have the given `id`
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: An error occured when fetching plants.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
    put:
      description: Edit a plant
      tags:
        - Plant
      responses:
        "200":
          description: The plant new details
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Plant"
        "404":
          description: No plant have the given `id`
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Another plant already have the same `name` and `latin_name`
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: An unexpected error happened.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        description: Plant to edit
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/PlantPayload"
  /plants/{id}/qrcode:
    parameters:
      - name: id
        in: path
        description: id of the plant
        required: true
        schema:
          type: string
    get:
      tags:
        - Plant
        - QRCode
      description: Get QRCode of a plant
      responses:
        "200":
          description: The QRCode
          content:
            "image/png":
              schema:
                type: string
                format: binary
        "404":
          description: No plant have the given `id`
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: An error occured when fetching plants.
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"